+++ yarn-project/end-to-end/src/fixtures/snapshot_manager.ts
